Rename code-sample module to samples
