+++ src/lib/samples/php.ts
@@ -2,7 +2,7 @@ import { snakeCase } from 'change-case'
 
 import type { Json, NonNullJson } from 'lib/json.js'
 
-import type { CodeSampleDefinition, Context } from './schema.js'
+import type { CodeSampleDefinition, Context } from './code-sample.js'
 
 export const createPhpRequest = (
   { request }: CodeSampleDefinition,
